sync node backend to java backend question upload

+++ backend-spring/exam-question/src/main/java/com/exam/question/entity/QuestionNotesEntity.java
@@ -23,7 +23,7 @@ public class QuestionNotesEntity {
     private String examCode;
 
     @Column(name = "question_id", nullable = false)
-    private Long questionId;
+    private Integer questionId;
 
     @Column(name = "note", columnDefinition = "TEXT")
     private String note;
